Simplify line distribution
Previous algorithm was complex and not guaranteed to work in every
case, necessitating a fallback.

We now use the simpler technique of spreading all texts equally along a
combined line (with first and last at the ends), then partitioning that
line and its texts by individual line extents.

The number of texts placed will always be equal to the total number
of texts.

TODO: Extents of texts are not considered here.  This is probably OK
in most cases.

Other changes:
 - use append instead of append!
 - add explanatory comments