Better integration with existing functionality
This commit is an extensive rewrite which creates the text spanner
stencil through an enhanced version of ly:line-spanner::print, which
has been translated from C++ to Scheme.

Previously, we calculated default stencils solely to gather position
and length information, and then built new stencils.  Here we opt for
a more integrated approach which builds a stencil with a single call
to a print function based on the one in the codebase.

The confusing system of line markers is simplified through use
of existing functionality: TextSpanner.bound-details.left-broken.text
and right-broken.text.

Additionally, beginning and ending texts are created through
bound-details.left.text and right.text.

The property 'text-spanner-line-count has been renamed
'inner-text-count-by-line to reflect a change.  No longer are the
beginning and end texts counted: the count is of inner texts only.
Thus, 0 for the first line means that no inner text is present, though
an initial text (left.text) will be.

The process of determining distribution of inner texts over available
lines has been simplified.

All of the previous functionality is present, except for collision
detection.  It may be reincorporated in a future commit.